+++ release/bar-chart/bar-horizontal-2d.component.d.ts
@@ -18,6 +18,8 @@ export declare class BarHorizontal2DComponent extends BaseChartComponent {
     schemeType: string;
     xAxisTickFormatting: any;
     yAxisTickFormatting: any;
+    xAxisTicks: any[];
+    yAxisTicks: any[];
     groupPadding: number;
     barPadding: number;
     roundDomains: boolean;