+++ release/line-chart/line-chart.component.d.ts
@@ -21,6 +21,8 @@ export declare class LineChartComponent extends BaseChartComponent {
     rangeFillOpacity: number;
     xAxisTickFormatting: any;
     yAxisTickFormatting: any;
+    xAxisTicks: any[];
+    yAxisTicks: any[];
     roundDomains: boolean;
     tooltipDisabled: boolean;
     showRefLines: boolean;